# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Nemo V2
  version: 1.0.0
...
paths:
...
  /api/v1/platform/account/{accountid}/subscription/history:
    get:
      tags:
        - FMS JS > FMS > AccountMangement > Subscription
      summary: Get Account Subscription History
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/account/{accountid}/credits/history:
    get:
      tags:
        - FMS JS > Platform > Accounts > Credits
      summary: Get Account Credits History
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/fms/historydata/vehicle/{vinno}/gps:
    post:
      tags:
        - FMS JS > FMS > History Data
      summary: Get GPS History Data
      description: Retrieve GPS history data for a specific vehicle within a time range
      security:
        - cookieAuth: []
      parameters:
        - name: vinno
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
          example: "MKMDJD7Y732282"
          description: "Vehicle VIN number"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["starttime", "endtime"]
              properties:
                starttime:
                  type: number
                  description: "Start time in milliseconds (timestamp)"
                  example: 1716150000000
                endtime:
                  type: number
                  description: "End time in milliseconds (timestamp)"
                  example: 1716236400000
      responses:
        "200":
          description: "GPS history data retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: "null"
                  data:
                    type: array
                    description: "Array of GPS data points"
                  msg:
                    example: "GPS history data listed successfully"
        "400":
          description: "Bad Request - Invalid input parameters"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: object
                  data:
                    type: "null"
                  msg:
                    type: string
        "403":
          description: "Forbidden - Authentication required"
        "500":
          description: "Internal Server Error"
  /api/v1/fms/historydata/vehicle/{vinno}/can:
    post:
      tags:
        - FMS JS > FMS > History Data
      summary: Get CAN History Data
      description: Retrieve CAN bus history data for a specific vehicle within a time range
      security:
        - cookieAuth: []
      parameters:
        - name: vinno
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
          example: "MKMDJD7Y732282"
          description: "Vehicle VIN number"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["starttime", "endtime"]
              properties:
                starttime:
                  type: number
                  description: "Start time in milliseconds (timestamp)"
                  example: 1716150000000
                endtime:
                  type: number
                  description: "End time in milliseconds (timestamp)"
                  example: 1716236400000
                canparams:
                  type: array
                  items:
                    type: string
                    maxLength: 128
                  minItems: 1
                  description: "Optional array of specific CAN parameters to retrieve"
                  example:
                    [
                      "intime",
                      "valid",
                      "tstamp",
                      "cycletime",
                      "charger_sts_live",
                      "bms_charger_plug_in_sts",
                    ]
      responses:
        "200":
          description: "CAN history data retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: "null"
                  data:
                    type: array
                    description: "Array of CAN data points"
                  msg:
                    type: string
                    example: "CAN history data listed successfully"
        "400":
          description: "Bad Request - Invalid input parameters"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: object
                  data:
                    type: "null"
                  msg:
                    type: string
        "403":
          description: "Forbidden - Authentication required"
        "500":
          description: "Internal Server Error"
  /api/v1/fms/historydata/vehicle/{vinno}/cangps:
    post:
      tags:
        - FMS JS > FMS > History Data
      summary: Get Merged CAN+GPS History Data
      description: Retrieve merged CAN and GPS history data for a specific vehicle within a time range
      security:
        - cookieAuth: []
      parameters:
        - name: vinno
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
          example: "MKMDJD7Y732282"
          description: "Vehicle VIN number"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["starttime", "endtime"]
              properties:
                starttime:
                  type: number
                  description: "Start time in milliseconds (timestamp)"
                  example: 1716150000000
                endtime:
                  type: number
                  description: "End time in milliseconds (timestamp)"
                  example: 1716236400000
                canparams:
                  type: array
                  items:
                    type: string
                    maxLength: 128
                  minItems: 1
                  description: "Optional array of specific CAN parameters to retrieve"
                  example:
                    [
                      "intime",
                      "valid",
                      "tstamp",
                      "cycletime",
                      "charger_sts_live",
                      "bms_charger_plug_in_sts",
                    ]
      responses:
        "200":
          description: "Merged CAN+GPS history data retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: "null"
                  data:
                    type: array
                    description: "Array of merged CAN+GPS data points"
                    items:
                      type: object
                      description: "Individual data point with merged CAN and GPS information"
                  msg:
                    type: string
                    example: "Merged CAN+GPS history data listed successfully"
        "400":
          description: "Bad Request - Invalid input parameters"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: object
                        description: "Validation error details"
                  data:
                    type: "null"
                  msg:
                    type: string
                    example: "Invalid input data"
        "403":
          description: "Forbidden - Authentication required"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "TOKEN_REQUIRED"
                  data:
                    type: "null"
                  msg:
                    type: string
                    example: "Account token is required"
        "500":
          description: "Internal Server Error - Failed to retrieve data"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INTERNAL_ERROR"
                  data:
                    type: "null"
                  msg:
                    type: string
                    example: "Failed to get merged CAN+GPS history data"
  /api/v1/fms/historydata/vehicle/latestdata:
    post:
      tags:
        - FMS JS > FMS > History Data
      summary: Get Vehicle Latest Data
      description: Retrieve the latest GPS and CAN data for multiple vehicles
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: ["vinnos"]
              properties:
                vinnos:
                  type: array
                  items:
                    type: string
                    maxLength: 128
                  minItems: 1
                  description: "Array of VIN numbers for which to retrieve latest data"
                  example:
                    ["MKMDJD7Y732282", "MKMDJD7Y732283", "MKMDJD7Y732284"]
      responses:
        "200":
          description: "Vehicle latest data retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: "null"
                  data:
                    type: object
                    properties:
                      gpsdata:
                        type: object
                        description: "Latest GPS data for each vehicle"
                      candata:
                        type: object
                        description: "Latest CAN data for each vehicle"
                  msg:
                    type: string
                    example: "Vehicle latest data fetched successfully"
        "400":
          description: "Bad Request - Invalid input parameters"
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: object
                    properties:
                      errcode:
                        type: string
                        example: "INPUT_ERROR"
                      errdata:
                        type: object
                  data:
                    type: "null"
                  msg:
                    type: string
        "403":
          description: "Forbidden - Authentication required"
        "500":
          description: "Internal Server Error"
...
  /api/v1/fms/account/subscription/history:
    get:
      tags:
        - FMS JS > FMS > AccountMangement > Subscription
      summary: Get Account Subscription History
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json: {}
...